Remade the clock design and code
Remade the clock as an iteration of Maeda's 8th clock which, whilst
still an updated design, is a much closer recreation of the original.

This version is a slightly modernised version of Maeda's clock as shown
through the clean, unpixelated, numbers. It also takes the changing of
size and shade and matches these up to 'ticking' of each second. This is
used as a non-written method of showing the passage of seconds as most
clocks do when they 'tick'.