Design and coding updates
Added the background changing effect - this now changes color to match
changes in times. Also added comments and rearranged some code.

Also made changes to debug.html, debug.js, and sketch.js to add in
functionality to debig changes in the day, and later the month and year.
This has been curtailed as I don't want to change the api functionality
too much.