Completed functionality
This is a final pass at functionality. Alarm is fully working as well as
all other features.

Also edited the base api and debug page and function to enable debugging
the new day, month, and year units.

Also added a possibility of two alarm countdowns and signal designs
which are picked at random.

Updated Readme, preview, and thumbnail to match.